Remove `benchmarking.rs`
